% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Location of Motor Vehicles Traffic Crashes}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{Year of interest}
}
\value{
The function plots a map of the state with points representing
locations at which the accidents occurred.
}
\description{
The function plots a map that shows where in the specified state the accidents
occurred during the specified year.
}
\details{
Internally, the function uses \code{\link{make_filename}} and
\code{\link{fars_read}} functions to load the data, filters by state and plots
accidents using the \code{\link{map}} function of the maps package. Only one state
and one year can be specified in the function call. If the user enters wrong
state number, the function will abort, throwing out an error. If the user filters
out all of the data, the function will print an appropriate message.
}
\examples{
\dontrun{
fars_map_state(10, 2014)
}

}
