% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import into R ".csv.bz2" Data from US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name (character string) of the file to be imported}
}
\value{
The function returns a tibble (a tidyverse version of the dataframe)
with the data of interest.
}
\description{
The function loads into R bz2-compressed CSV for the year of interest. Canonical
filename can be conveniently constructed with the function \code{\link{make_filename}}.
}
\details{
The function uses readr::read_csv to import the file
and dplyr::tbl_df to construct the dataframe. It will not show
how it formatted each column nor its loading progress. If it fails to find a
specified file, it will throw an error, notifying the user that the file does
not exist. The function assumes that the file is in the current working directory.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars_read(make_filename(2013))
}

}
