% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Import Into R Several Years of Data, Retain Only Month and Year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an integer vector specifying which years should be imported into R}
}
\value{
The function returns a list of tibbles. Each of the tibbles has only
two columns, MONTH and year.
}
\description{
This function imports several years of data and retains from it only the month
and year columns.
}
\details{
Internally, the function uses \code{\link{make_filename}} function to
generate the filenames and then \code{\link{fars_read}} function to import the data.
If the data for the specified year are not available, the function will create a
list entry with NULL value.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
fars_read_years(c(2013, 2017, 2014))
}

}
