% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise Number of Accidents Per Months}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer vector specifying which years should be imported into R}
}
\value{
The function returns a tibble with the number of accidents that occurred
in a specific month of a specified year. The months (1-12) are shown in
rows and the years in columns.
}
\description{
The function counts the number of accidents that occurred in each month of
the specified years.
}
\details{
Internally, the function builds on the function \code{\link{fars_read_years}}
and summarises the data by counting the number of month-year occurrences.
}
\examples{
fars_summarize_years(c(2013, 2014))

}
