% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Infer Filename From Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string of four numbers that denote the year of interest.}
}
\value{
This function returns a character string which corresponds to the
canonical name of a file that contains bz2-compressed comma-separated values
for the year of interest.
}
\description{
The function generates the canonical filename for the year of interest.
}
\note{
Entered year is first coerced into integer and only then used to generate
a filename. Entering non-numbers will coerce entered string into NA.
}
\examples{
make_filename(2013)
make_filename("2013")

}
